pyproject.toml
[project]
name = "chainlit-agent-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.8.3",
    "dotenv>=0.9.9",
    "openai-agents>=0.3.3",
    "rich>=14.2.0",
]

[project.scripts]
chainlit-agent-assignment = "chainlit_agent_assignment:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
